Fix Tenma72_13360 detection
Currently, Tenma72_13360 is not detected by the
instantiate_tenma_class_from_device_response and returns
the default Tenma72_2545 which is not compatible with Tenma72_13360.

In fact, instantiate_tenma_class_from_device_response parses all
subclasses of Tenma72Base. However, Tenma72_13360 is a subclass of
Tenma72_13360_base.

Therefore, we need:
  - A common interface which involves changing the interface of
    Tenma72_13360 to add channel (that will not be used because
    Tenma72_13360 has only one channel).
  - Or an object that redirects calls to the right implementation

The second solution was implemented in this patch to avoid changing the
interface of Tenma72_13360 and probably break user's software.
This patch adds a new class called TenmaPowerSupply that will create
a Tenma device based on the response of the device on the serial port.
If the user calls "setVoltage(channel, mV)" and the device is Tenma72_13360
the call will be redirected to Tenma72_13360.setVoltage(mV).
The class Tenma72_13360 remains unchanged. Therefore, user can call
directly Tenma72_13360.setVoltage(mV).

Tested on HW Tenma72-13360 1xChannel 60V 15A using tenmaControl.py
  - setCurrent
  - setVoltage
  - getCurrent
  - getVoltage