Fixes #20, by providing a common interface to tenma-control for 13360. (#23)
The original problem comes from Tenma72 family and the 13360 variant do
not share a common signature for functions `setCurrent(channel,mv)` and
`setCurrent(mv)` since there is no channel.

Ideally I'd like to simply force the same signatures, but this would
break user-code , forcing us to a 2.x version. We might do it in the
future but not now.

This change provides an Adapter class wrapper that we'll use for the
13360 family (and in case any other family comes with single channel).
`instantiate_tenma_class_from_device_response` will take this into
account and return the Adapter class to be used by `tenmaControl`.